+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Added `MutableSubscriptionSet.removeByType` for removing subscriptions by their realm object type. (Issue [#317](https://github.com/realm/realm-dart/issues/317))
 
 ### Fixed
-* None
+* Fixed a wrong mapping for `AuthProviderType` returned by `User.provider` for google, facebook and apple credentials.
 
 ### Compatibility
 * Realm Studio: 12.0.0 or later.